requirements.txt
addict==2.4.0
albumentations==1.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
einops==0.6.1
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
Markdown==3.4.3
markdown-it-py==3.0.0
matplotlib==3.7.1
mdurl==0.1.2
mmcv==2.0.1
mmengine==0.7.4
model-index==0.1.11
networkx==3.1
numpy==1.24.4
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.8.0
prettytable==3.8.0
pycryptodome==3.18.0
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tomli==2.0.1
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
tqdm==4.65.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
yapf==0.40.1
zipp==3.15.0
